Remove PL011 initialization from bootloader
