+++ BoardSupport/qemu-aarch64/BoardDefs.hpp
@@ -1,0 +1,11 @@
+#ifndef _BOARD_DEFS_HPP_
+#define _BOARD_DEFS_HPP_
+
+typedef enum DeviceToAddress
+{
+	DEVICE_TYPE_UART = 0xFFFFFFF7FFDF0000,
+	DEVICE_TYPE_GIC_DIST = 0xFFFFFFF7FFDF2000,
+	DEVICE_TYPE_GIC_CPU = 0xFFFFFFF7FFDF4000,
+}DeviceToAddress;
+
+#endif
