+++ BoardSupport/qemu-aarch64/BoardInit/Board.cpp
@@ -1,9 +1,8 @@
 #include <kern_runtime.hpp>
 #include <kern_datatypes.hpp>
-#include <BoardSupport/qemu-aarch64/Board.hpp>
 
-/* On qemu-virt, we definitely have a uart. */
-#include <Devices/PL011/PL011.hpp>
+#include <BoardDefs.hpp>
+#include <BoardSupport/qemu-aarch64/BoardInit/Board.hpp>
 
 #include <arch/aarch64/arch.hpp>
 #include <arch/aarch64/gic.hpp>
@@ -11,34 +10,6 @@
 
 #include <arch/aarch64/vmm/vmm.hpp>
 
-typedef enum DeviceToAddress
-{
-	DEVICE_TYPE_UART = 0xFFFFFFF7FFDF0000,
-	DEVICE_TYPE_GIC_DIST = 0xFFFFFFF7FFDF2000,
-	DEVICE_TYPE_GIC_CPU = 0xFFFFFFF7FFDF4000,
-}DeviceToAddress;
-
-
-void WriteSerialChar(CHAR Char)
-{
-	pantheon::pl011::PL011WriteChar(0, Char);
-}
-
-void WriteString(const CHAR *String)
-{
-	UINT64 Index = 0;
-	while (String[Index])
-	{
-		CHAR CurChar = String[Index];
-		if (CurChar == '\n')
-		{
-			WriteSerialChar('\r');
-		}
-		WriteSerialChar(CurChar);
-
-		Index++;
-	}
-}
 
 extern "C" void BoardInit(pantheon::vmm::PageTable *TTBR1, pantheon::vmm::PageAllocator &PageAllocator)
 {
@@ -59,8 +30,6 @@ extern "C" void BoardInit(pantheon::vmm::PageTable *TTBR1, pantheon::vmm::PageAl
 	PageAllocator.MapLower(TTBR1, DEVICE_TYPE_UART, 0x09000000, pantheon::vmm::BlockSize::L3BlockSize, DeviceMMIOEntry);
 	PageAllocator.MapLower(TTBR1, DEVICE_TYPE_GIC_DIST, 0x08000000, pantheon::vmm::BlockSize::L3BlockSize, DeviceMMIOEntry);
 	PageAllocator.MapLower(TTBR1, DEVICE_TYPE_GIC_CPU, 0x08010000, pantheon::vmm::BlockSize::L3BlockSize, DeviceMMIOEntry);
-
-	pantheon::pl011::PL011Init(DEVICE_TYPE_UART, 0);
 
 	pantheon::arm::GICSetMMIOAddr(pantheon::arm::GIC_CLASS_DISTRIBUTOR, DEVICE_TYPE_GIC_DIST);
 	pantheon::arm::GICSetMMIOAddr(pantheon::arm::GIC_CLASS_CPU_INTERFACE, DEVICE_TYPE_GIC_CPU);
