+++ BoardSupport/qemu-aarch64/BoardInit/Board.hpp
@@ -3,6 +3,6 @@
 #ifndef _BOARD_HPP_
 #define _BOARD_HPP_
 
-#define PUTCHAR_FUNC(x) WriteSerialChar(x)
+/* Not in use for now */
 
 #endif
