+++ BoardSupport/qemu-aarch64/BoardRuntime/CMakeLists.txt
@@ -1,0 +1,6 @@
+LIST(APPEND BOARDRT_HEADERS BoardRT.hpp)
+LIST(APPEND BOARDRT_SOURCES BoardRT.cpp)
+
+ADD_LIBRARY(BoardRT STATIC
+	${BOARDRT_HEADERS}
+	${BOARDRT_SOURCES})
