+++ BoardSupport/qemu-aarch64/Boot/Boot.cpp
@@ -378,66 +378,6 @@ void Initialize(fdt_header *dtb, pantheon::vmm::PageAllocator &Allocator)
 	}
 }
 
-void PrintDTB(fdt_header *dtb)
-{
-	volatile bool CheckMe = CheckHeader(dtb);
-	if (!CheckMe)
-	{
-		/* Loop forever: can't really do anything. */
-		for (;;) {}
-	}
-
-	DeviceTreeBlob DTBState(dtb);
-	CHAR CurDevNode[512];
-	ClearBuffer(CurDevNode, 512);
-
-	while (!DTBState.EndStruct())
-	{
-		DTBState.NextStruct();
-		FDTNodeType CurNode = DTBState.GetStructType();
-		if (CurNode == FDT_PROP)
-		{
-			UINT64 Offset = DTBState.GetPropStructNameIndex();
-			CHAR Buffer[512];
-			ClearBuffer(Buffer, 512);
-			DTBState.CopyStringFromOffset(Offset, Buffer, 512);
-			SERIAL_LOG_UNSAFE("%s%s%s", CurDevNode, " : ", Buffer);
-			if (IsStringPropType(Buffer) || IsStringListPropType(Buffer))
-			{
-				CHAR Buffer2[512];
-				ClearBuffer(Buffer2, 512);
-				DTBState.CopyStringFromStructPropNode(Buffer2, 512);
-				SERIAL_LOG_UNSAFE("%s%s%s", " (", Buffer2, ")");
-			}
-			else if (IsU32PropType(Buffer))
-			{
-				UINT32 U32;
-				DTBState.CopyU32FromStructPropNode(&U32);
-				SERIAL_LOG_UNSAFE("%s%u%s", " (", U32, ")");
-			}
-			else if (IsU64PropType(Buffer))
-			{
-				UINT64 U64;
-				DTBState.CopyU64FromStructPropNode(&U64);
-				SERIAL_LOG_UNSAFE("%s%u%s", " (", U64, ")");
-			}
-			
-			CHAR DevName[512];
-			UINT64 Addr;
-			DTBState.NodeNameToAddress(CurDevNode, DevName, 512, &Addr);
-			SERIAL_LOG_UNSAFE("%s", "\n");
-		}
-		else if (CurNode == FDT_BEGIN_NODE)
-		{
-			ClearBuffer(CurDevNode, 512);
-			DTBState.CopyStringFromStructBeginNode(CurDevNode, 512);
-			SERIAL_LOG_UNSAFE("%s%s%s", "<<", CurDevNode, ">>\n");
-
-		}
-	}
-	SERIAL_LOG_UNSAFE("%s\n", "finished going through dtb");
-}
-
 static pantheon::vmm::PageAllocator InitialPageTables;
 
 static void SetupPageTables()
@@ -552,7 +492,6 @@ extern "C" InitialBootInfo *BootInit(fdt_header *dtb, void *initial_load_addr, v
 		 */
 		BoardInit(&TTBR1, InitialPageTables);
 		Initialize(dtb, InitialPageTables);
-		PrintDTB(dtb);
 
 		UINT64 InitAddr = (UINT64)initial_load_addr;
