+++ CMakeLists.txt
@@ -94,6 +94,7 @@ TARGET_LINK_LIBRARIES(pantheon Board)
 TARGET_LINK_LIBRARIES(pantheon Printf)
 TARGET_LINK_LIBRARIES(pantheon System)
 TARGET_LINK_LIBRARIES(pantheon Boot)
+TARGET_LINK_LIBRARIES(pantheon BoardRT)
 
 ADD_CUSTOM_COMMAND(TARGET ${CMAKE_PROJECT_NAME} POST_BUILD
 	COMMAND ${CMAKE_OBJCOPY} -O binary ${CMAKE_PROJECT_NAME} kernel.img)
